Add devices auto refresh & progress indicator
